+++ demisto_sdk/commands/validate_poc/validators/id_name_validator.py
@@ -1,5 +1,3 @@
-from typing import Tuple, Type
-from demisto_sdk.commands.content_graph.objects.base_content import BaseContent
 from demisto_sdk.commands.content_graph.objects.classifier import Classifier
 from demisto_sdk.commands.content_graph.objects.content_item import ContentItem
 from demisto_sdk.commands.content_graph.objects.integration import Integration
